\name{salvimat_qqnorm}
\alias{salvimat_qqnorm}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Simulate qqnorm plot
}
\description{
Simulate qqnorm plot
}
\usage{
salvimat_qqnorm(lmModel, n_sim = 500)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{lmModel}{
A lm-Model
}
  \item{n_sim}{
number of simulations
}
}
\keyword{ qqnorm }

