\name{salvimat_ta}
\alias{salvimat_ta}
\title{
Simulate TA Plot
}
\description{
Simulate TA Plot
}
\usage{
salvimat_ta(lmModel, method = "LS", n_sim = 500)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{lmModel}{
an lm or rlm Model
}
  \item{method}{
3 different method, "LS" , "M" or "MM"
}
  \item{n_sim}{
number of simulation
}
}
