\name{StMoSim-package}
\alias{StMoSim-package}
\alias{StMoSim}
\docType{package}
\title{
Simulate QQ-Norm and TA-Plot

}
\description{
With this package you can simulate several lines into the QQnorm-Plot and the TA-plot under Gaussian distribution
}
\details{
\tabular{ll}{
Package: \tab StMoSim\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2010-05-01\cr
License: \tab none\cr
LazyLoad: \tab yes\cr
}
2 functions
salvimat_qqnorm and salvimat_ta
}
\author{
Matthias Salvisberg

Maintainer: Matthias Salvisberg <matthias.salvisberg@gmail.com>
}

\keyword{ package }

