\name{StMoSim-package}
\alias{StMoSim-package}
\alias{StMoSim}
\docType{package}
\title{
Simulation in the QQ-Norm-Plot

}
\description{
With this package you can simulate several lines into the QQ-Norm-Plot under the assumption of Gaussian distribution.
This can be very useful in residual analysis if you have to evaluate whether a model fits the data or not.
}
\details{
\tabular{ll}{
Package: \tab StMoSim\cr
Type: \tab Package\cr
Version: \tab 2.0\cr
Date: \tab 2011-03-31\cr
License: \tab none\cr
LazyLoad: \tab yes\cr
}
function:
qqnormSim
}
\author{
Matthias Salvisberg

Maintainer: Matthias Salvisberg <matthias.salvisberg@gmail.com>
}

\keyword{ package }

